backend/requirements.txt
functions-framework==3.*
google-cloud-firestore>=2.11.0